+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "config": "ts-node ./scripts/setenv.ts",
+    "config": "ts-node ./src/assets/setenv.ts",
     "prestart": "npm run config -- --environment=dev",
     "start": "ng serve",
     "build": "npm run config -- --environment=prod && ng build --prod",
@@ -58,6 +58,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
+    "typescript": "~4.1.5",
+    "yargs": "^16.2.0"
   }
 }
