--- src/environments/environment.prod.ts
@@ -1,5 +1,0 @@
-export const environment = {
-    production: true,
-    apiUrl: '',
-    API_KEY: 'SECRET'
-  };
